Deleted some extraneous files.
Streamlined the loading and saving of zones and players. There is no issue with pickling code objects (i.e., logic modules), and we now don't have to worry about pickling the references to zone/rooms when saving a player object. The MUD now translates it to an integer upon saving, and back to a reference upon loading.