Okay!  Added two commands: Search and Commands.
Search allows a player to search the template database for a template by name (for now).  This will be expanded in the future to allow searching for any statistic.

Commands allows a player to receive a comma-separated list of the commands that they have available.

Other files have changed or been modified to suit the ends of these two commands (and other buggy type stuff I found along the way). :)