Okay, couple of changes today.
First, the commands command is updated to use the new columnizer utility which is also new and will allow us to turn a list of strings into a nicely formatted display by calling a single command. I also wrote this up and added it this morning although it is based upon Kuros' code for the original SlitherMUD.

I did preliminary work on making wizview feasible by altering the way look worked in basic room logic.  There will be more changes to come, but I'll get to them tomorrow most likely.

-V